Migration visualisationConfigs mapping to useVisualisationConfigs
